requirements.txt
numpy==2.2.2
pillow==11.1.0
torch==2.6.0
torchvision==0.21.0
tqdm
